requirements.txt
Flask
flask-mysqldb
python-dotenv
pytest
pytest-mock
coverage